Organize Presentation.kt
